Add functions for mutably building an Nfa.

+++ src/regex.rs
@@ -37,7 +37,7 @@ impl Regex {
     /// Converts this Regex to an Nfa using Thompson's construction.
     ///
     /// The states are `u16`, and the alphabet (Σ) is the set of `char` values.
-    pub fn to_nfa(&self) -> Nfa<u16, u8> {
+    pub fn to_nfa(&self) -> Nfa<u64, u8> {
         panic!("Not implemented");
     }
